(dev): presentation layer

+++ app/src/main/java/vn/hoanguyen/cleanarchitecture/trivianumber/app/features/trivia/data/datasources/NumberTriviaLocalDataSource.kt
@@ -10,7 +10,7 @@ interface NumberTriviaLocalDataSource {
      * Gets the cached [NumberTriviaDto] which was gotten the last time the user has an internet connection
      * Throws a [CachedException] if no cached data is present
      */
-    suspend fun getConcreteNumberTrivia(number: Double): NumberTriviaDto
+    suspend fun getConcreteNumberTrivia(number: Int): NumberTriviaDto
 
     /**
      * Call the api http://numbersapi.com/random/trivia?json
