+++ app/src/main/java/vn/hoanguyen/cleanarchitecture/trivianumber/app/features/trivia/domain/repositories/NumberTriviaRepository.kt
@@ -8,6 +8,6 @@ import vn.hoanguyen.cleanarchitecture.trivianumber.core.functional.Either
  * Created by Hoa Nguyen on Aug 06 2021.
  */
 interface NumberTriviaRepository {
-    suspend fun getConcreteNumberTrivia(number: Double): Either<IFailure, NumberTrivia>
+    suspend fun getConcreteNumberTrivia(number: Int): Either<IFailure, NumberTrivia>
     suspend fun getRandomNumberTrivia(): Either<IFailure, NumberTrivia>
 }
