+++ app/src/test/java/vn/hoanguyen/cleanarchitecture/trivianumber/app/features/trivia/domain/usecases/UseCaseGetConcreteNumberTriviaTest.kt
@@ -34,10 +34,10 @@ internal class UseCaseGetConcreteNumberTriviaTest {
     fun `should return numberTrivia success from the repository`() =
         runBlockingTest {
             //Arrange
-            val number = 1.0
+            val number = 1
             val numberTrivia = NumberTrivia(
                 text = "SomeText",
-                number = number,
+                number = number.toDouble(),
                 found = true,
                 type = "trivia"
             )
@@ -59,7 +59,7 @@ internal class UseCaseGetConcreteNumberTriviaTest {
     fun `should return Failure from the repository`() =
         runBlockingTest {
             //Arrange
-            val number = 1.0
+            val number = 1
             val failure = FailureNumberTriviaNotFound()
             coEvery { repository.getConcreteNumberTrivia(number) } returns Either.Left(
                 failure
